refactor(guards): Move anwaltlicheVertretung guards into own file
